Ajaxify submissions
Link submissions are now handle through an ajax call. Altered the
submission UI to show a loading spinner why the backend is querying for
the name of the video. An update view of the queue is then returned upon
a successful submission.

yt_controller now makes new connections each time it contacts yt_player.
Short of framing and addressing the messages, it would probably have
lead to collisions in messages if more users are connected if I kept it
one maintained connection.