Thread all incoming requests
All RPC messages that yt_player receives are now placed inside the job
queue to be handled by a thread. The hope is to be a little more
responsive.

Fixes a couple ajax errors. The now playing request was unnecessarily
re-registering callbacks and causing a loop of callbacks to occur.
Adding a new song wasn't re-registering the callback on the queue title.