Tweaks to the behavior of the songs in the queue
Added a visual hint that a song in the queue can be tapped on to expand
the detail view.

Made the whole row tappable.

Fix bug where the detail view wasn't expanding in Safari.

Changed the youtube link in the detail view to be a part of a button
group in anticipation of further changes down the line.