Tweak the "now playing" banner
Moved the anchor tag on the title of the currently playing song to a
link that's included in the hidden detail view. This makes the banner an
easier target for tapping to expand the banner without accidentally
hitting the link.

I changed the response message to an "add" to just say succcess instead
of return the new state of the queue. In behavior.js, it now requests
the now playing and queue after receiving a response to add. This was to
add a way for the page to auto-refresh after a user submits a song. This
method also give you a more update view of the state so if a user
submits the first song, it usually goes straight to the now playing
banner instead of being shown in the queue.

I also fixed the json message extractor to handle curly brackets within
json messages. However, I don't think this is necessary anymore each RPC
establishes a new connection and tears it down on completing. Therefore,
there should only ever be one json message read of the pipe.