Enable player config and fix cookies
Added a configuration value enable the player on the server side. This
is useful for time when you're working with items in the queue and don't
want anything to get pulled out.

Moved the session cookie over to a permanent cookie that expires after
one day. On my phone at least, when the page is revived by my browser,
the session was getting cleared. This appears to fix the problem.