Change to user ids
New users are now assigned a user id that unique to each unique user
name. This user id is what is used to identify users throughout the
system.

No two people with the same name can be logged in at the same time.

Various rpc messages were updated to return a status and error message.

Add hooks into a sqlite database on the backend. Right now, it's used to
keep track of the usernames and their ids.

Add logo for the site.