Change to user ids
Each user session is now assigned a unique user id that is used
throughout the system to identify a particular user.

Various rpc messages were updated to return a status and error message.

Add hooks into a sqlite database on the backend. Right now, it's used to
keep track of the users and submissions. There's a settings table that
could possibly be used in the future to store settings.

Add logo for the site.

Smoothed out the transitions on the now playing banner.

The session cookie now lasts for one year in order to allow return users
to reclaim their id as long as they still have their cookie.