Add error reporting to webpage and other fixes
A couple of the RPC messages have been updated to return an alert object
and status code. This is currently being used to surface errors on the
webpage in the form a little banner below the Now Playing banner. When
the server throws and error, it will set the http status code to 500 and
the javascript should respond accordingly. One could also use this to
print other messages from the server so don't abuse this.

When the javascript fails to contact the server or retrieve an error, it
will fallback to creating a popup to alert the user.

The login page also uses this style of error reporting now.

Added comments to the scheduler classes.

Fixed a bug in the removal code in the round robin class. It will now
only shift a user's videos forward a round if the video was in a later
round than the one removed.