More database work
A user's logged in state is now kept track of inside the users table.
The column is there in case we ever want to allow two browsers to be
logged in with the same user id.

The session cookie now lasts for one year.

The logout option now only allows a user to change his/her display name.
The user id will follow the user's browser for as along as the session
cookie is valid.

Add tables for submissions and user settings.